pyproject.toml
[tool.poetry]
name = "fast_tut"
version = "0.1.0"
description = ""
authors = ["Ryan Jones <ryan.jones@disneystreaming.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
SQLAlchemy = "^1.4.32"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
uvicorn = "^0.17.5"
psycopg2-binary = "^2.9.3"
pydantic = "^1.9.0"
email-validator = "^1.1.3"
python-jose = {extras = ["cyrptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
